// modules/chat/util/types.ts
export interface ChatMessage extends RawChatMessage {
  createdAt: Date;
}
// ...
export interface ChatEventBase {
  type: string;
  data?: unknown;
}

export type ChatEventHandler<T = OutboundChatEvent> = (
  event: ChatEventBase,
  connection: ChatConnection,
) => MaybePromise<T | null>;

export type HandlerConfig = {
  handler: ChatEventHandler;
  queue?: "msg";
};

// =====================
// Inbound Events
// =====================

export type InboundChatEvent =
  | NewChatMsgEvent
  | EditedChatMsgEvent
  | DeletedChatMsgEvent
  | LastSeenFeedItemEvent
  | LoadOlderMessagesEvent
  | UserTypingEvent
  | SubscriberOnlineEvent;

export interface NewChatMsgEvent extends ChatEventBase {
  type: "new-chat-msg";
  data: Pick<ChatMessage, "text"> & { clientMsgId: string };
}

export interface EditedChatMsgEvent extends ChatEventBase {
  type: "edited-chat-msg";
  data: Pick<ChatMessage, "id" | "text">;
}

export interface DeletedChatMsgEvent extends ChatEventBase {
  type: "deleted-chat-msg";
  data: Pick<ChatMessage, "id">;
}

export interface LastSeenFeedItemEvent extends ChatEventBase {
  type: "last-seen-feed-item";
  data: { id: string };
}

export interface LoadOlderMessagesEvent extends ChatEventBase {
  type: "load-older-messages";
  data: { olderMsgsCursor: string };
}

export interface UserTypingEvent extends ChatEventBase {
  type: "user-typing";
  data: {
    username: string;
    lastUserMsgId?: string;
  };
}

export interface SubscriberOnlineEvent extends ChatEventBase {
  type: "subscriber-online";
  data: {
    subscriberId: string;
    skipChatSubUpdate?: boolean;
  };
}

// =====================
// Outbound Events
// =====================

export type OutboundChatEvent =
  | ChatReadyEvent
  | NewChatMsgEventResp
  | EditedChatMsgEventResp
  | DeletedChatMsgEventResp
  | LoadOlderMessagesEventResp
  | UserTypingEvent
  | FeedebEvent
  | ErrorEvent;